bot-server/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-telegram-bot==20.7
httpx==0.26.0
pydantic-settings==2.1.0
